added global context, and take location from AlarmReceiver itself
we first attempted to call onLocation from LocationUpdatesService, but realise that while it works on Huawei, it didnt work on Xiaomi because the service is also killed even the app is swiped away. It seems that on the Xiaomi only the AlarmReceiver can run when the app is swiped away.