added firebase server timestamp
Firebase Realtime Database provides offline storage as well, and automatically syncs the offline data to the server when network connection is resumed.

This presents the possibility of our location reports only syncing in bulk to the server not in real time.

Adding the Server Timestamp allows us to compare the time between the app and server timestamp to see if there was a delay.